WOrking distance thingy

+++ include/ultrasonic_distance.h
@@ -11,14 +11,14 @@ typedef struct {
   gpio_num_t gpio_trigger;
   gpio_num_t gpio_echo;
   int16_t timeout_in_u_seconds;
-  int16_t measured_array[numberOfMeasurements];
-  int16_t average_measured;
+  long measured_array[numberOfMeasurements];
+  long average_measured;
   uint8_t task_done;
 } distance_measurements;
 
 void start_measurement(gpio_num_t gpio_pin_trig);
-int16_t wait_for_echo(gpio_num_t gpio_pin_echo, uint8_t timeout, uint8_t level);
-int16_t get_distance(int16_t u_seconds);
+long wait_for_echo(gpio_num_t gpio_pin_echo, int16_t timeout, uint8_t level);
+long get_distance(long u_seconds);
 esp_err_t init_distance_gpio(gpio_num_t gpio_trigger, gpio_num_t gpio_echo);
 void measure_distance_task(void *pvParameters);
